Ajusta tipo do campo step_number de bigint para int;
refs #1884

+++ db/migrate/20190118201914_adjust_function_steps_by_classroom_to_use_step_number.rb
@@ -7,7 +7,7 @@ class AdjustFunctionStepsByClassroomToUseStepNumber < ActiveRecord::Migration
         _classroom_id INT
       )
       RETURNS TABLE (
-        step_number BIGINT,
+        step_number INT,
         step_id INT,
         start_at DATE,
         end_at DATE,
