+++ db/migrate/20190118201915_adjust_function_step_by_classroom_by_date.rb
@@ -8,7 +8,7 @@ class AdjustFunctionStepByClassroomByDate < ActiveRecord::Migration
         _recorded_at DATE
       )
       RETURNS TABLE (
-        step_number BIGINT,
+        step_number INT,
         step_id INT,
         start_at DATE,
         end_at DATE,
