+++ db/migrate/20190118201916_create_function_step_by_classroom_by_step_number.rb
@@ -6,7 +6,7 @@ class CreateFunctionStepByClassroomByStepNumber < ActiveRecord::Migration
         _step_number INT
       )
       RETURNS TABLE (
-        step_number BIGINT,
+        step_number INT,
         step_id INT,
         start_at DATE,
         end_at DATE,
